+++ app/components/developers/query_sort_button_component.rb
@@ -4,6 +4,10 @@ module Developers
 
     delegate :sort, :search_query, to: :query
 
+    def render?
+      Feature.enabled?(:sort)
+    end
+
     def initialize(query:, user:, form_id:, scope: nil)
       @query = query
       @user = user