+++ test/components/developers/new_fields_component_test.rb
@@ -3,7 +3,7 @@ require "test_helper"
 module Developers
   class NewFieldsComponentTest < ViewComponent::TestCase
     test "renders if the user's developer profile is missing fields" do
-      developers(:one).update!(available_on: nil)
+      developers(:one).update!(scheduling_link: nil)
       render_inline NewFieldsComponent.new(users(:developer))
       assert_text I18n.t("developers.new_fields_component.title")
     end